% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_names.R
\name{add_names}
\alias{add_names}
\title{Function to add names to network for the user.}
\usage{
add_names(grn, gen_names)
}
\arguments{
\item{grn}{a GRN object from KBoost.}

\item{gen_names}{a vector with the gene names.}
}
\description{
Function to add names to network for the user.
}
