% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tf_get_post.R
\name{add_tf_get_post}
\alias{add_tf_get_post}
\title{Function to add a new tf to an existing model and calculate posteriors.}
\usage{
add_tf_get_post(priors, lliks, model, f, reg, kpca, TFs)
}
\arguments{
\item{priors}{a GxK numeric matrix with the log-priors of TF regulation models.}

\item{lliks}{a GxK numeric matrix with the log-marginal likelihoods of TF regulation models.}

\item{model}{a GxK binary matrix with the models at the previous iteration.}

\item{f}{an NxG matrix with the predicted gene expression values.}

\item{reg}{a list produced from the ort_reg function.}

\item{kpca}{a lits with the Kernel Principal Components.}

\item{TFs}{indexes of the genes in the system integers that are TFs.}
}
\description{
Function to add a new tf to an existing model and calculate posteriors.
}
