% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{A function to format a check that the user input to KBoost is correct.}
\usage{
check_input(X, TFs, g, v, prior_weights, ite)
}
\arguments{
\item{X}{an NxG numeric matrix with the expression values of G genes and N obersvations.}

\item{TFs}{a Kx1 numeric matrix with integers that are the indexes of columns of X with genes that encode TFs.}

\item{g}{a positive scalar with the width parameter for the RBF kernel. (default g = 40)}

\item{v}{a double between 0 and 1 with the shrinkage parameter. (default v =0.1)}

\item{prior_weights}{a GxK matrix with the prior weights for the prior probability that a TF regulates a gene.(default 0.5 for all values)}

\item{ite}{an integer with the number of iterations (default ite = 3)}
}
\value{
list with input revised and default values added
}
\description{
A function to format a check that the user input to KBoost is correct.
}
