% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D4_multi_1-data.R
\docType{data}
\name{D4_multi_1}
\alias{D4_multi_1}
\title{Dream 4 multifactorial pertubation challenge dataset 1}
\format{
matrix
}
\source{
https://www.synapse.org/#!Synapse:syn3049712/wiki/74628
}
\usage{
D4_multi_1
}
\description{
Each column is a gene and each row is a simulated experiment.
}
\examples{

data(D4_multi_1)
}
\references{
Marbach D, Prill RJ, Schaffter T, Mattiussi C, Floreano D, and Stolovitzky G. Revealing strengths and weaknesses of methods for gene network inference. PNAS, 107(14):6286-6291, 2010. Pubmed
}
\keyword{datasets}
