% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D5_3-data.R
\docType{data}
\name{D5_3}
\alias{D5_3}
\title{Dream 5 challenge dataset 3}
\format{
matrix
}
\source{
https://www.synapse.org/#!Synapse:syn2787209/wiki/70349
}
\usage{
D5_3
}
\description{
Each column is a gene and each row is an experiment.
}
\examples{

data(D5_3)
}
\references{
Marbach D, Costello J, Kuffner R, Vega N, Prill R,Camcaho D,Allison K, the DREAM5 Consortium, Kellis M, Collins J and Stolovitzky G. Wisdom of crowds for robust gene network inference. Nat Methods. 2012 Aug; 9(8): 796–804.
}
\keyword{datasets}
