% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D5_4-data.R
\docType{data}
\name{D5_4}
\alias{D5_4}
\title{Dream 5 challenge dataset 4}
\format{
matrix
}
\source{
https://www.synapse.org/#!Synapse:syn2787209/wiki/70349
}
\usage{
D5_4
}
\description{
Each column is a gene and each row is an experiment.
}
\examples{

data(D5_4)
}
\references{
Marbach D, Costello J, Kuffner R, Vega N, Prill R,Camacho D,Allison K, the DREAM5 Consortium, Kellis M, Collins J and Stolovitzky G. Wisdom of crowds for robust gene network inference. Nat Methods. 2012 Aug; 9(8): 796–804.
}
\keyword{datasets}
