% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d5_check.R
\name{d5_check}
\alias{d5_check}
\title{Function to produce the AUPR and AUROC Results on the DREAM4 Multifactorial Challenge.}
\usage{
d5_check(v, g, ite, write_res)
}
\arguments{
\item{v}{a number between 0 and 1 that is the shrinkage parameter}

\item{g}{a number larger than 0 that is the width parameter for the RBF Kernel}

\item{ite}{an integer with number of iterations.}

\item{write_res}{TRUE if the results are to be written as tables.}
}
\description{
Function to produce the AUPR and AUROC Results on the DREAM4 Multifactorial Challenge.
}
