% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_D4_multi_5-data.R
\docType{data}
\name{G_D4_multi_5}
\alias{G_D4_multi_5}
\title{Gold Standard Dream 4 multifactorial pertubation challenge dataset 5}
\format{
matrix
}
\source{
https://www.synapse.org/#!Synapse:syn3049712/wiki/74628
}
\usage{
G_D4_multi_5
}
\description{
Each column is a gene and each row is a simulated experiment.
}
\examples{

data(G_D4_multi_5)
}
\references{
Marbach D, Prill RJ, Schaffter T, Mattiussi C, Floreano D, and Stolovitzky G. Revealing strengths and weaknesses of methods for gene network inference. PNAS, 107(14):6286-6291, 2010. Pubmed
}
\keyword{datasets}
