% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gerstein_Prior_ENET_2-data.R
\docType{data}
\name{Gerstein_Prior_ENET_2}
\alias{Gerstein_Prior_ENET_2}
\title{Gene Regulatory Network from human ChIP-Seq data in Encode}
\format{
matrix
}
\usage{
Gerstein_Prior_ENET_2
}
\description{
A gene regulatory network inferred from the ChIp-Seq Encode dataset
Table with two columns

The fist column is a transcription factor and the second is a gene.
}
\examples{

data(Gerstein_Prior_ENET_2)
}
\references{
Gerstein, M.B., et al. Architecture of the human regulatory network derived from ENCODE data. Nature 2012;489(7414):91-100.
}
\keyword{datasets}
