% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_Gerstein.R
\name{get_prior_Gerstein}
\alias{get_prior_Gerstein}
\title{Function to build a prior automatically using a previously built Network on ChIP-Seq.}
\usage{
get_prior_Gerstein(gen_names, TFs, pos_weight, neg_weight)
}
\arguments{
\item{gen_names}{the gene names in Symbol nomenclature.}

\item{TFs}{the indexes of gene names which are TFs.}

\item{pos_weight}{the prior weight for edges that were previously found in the Gerstein network.}

\item{neg_weight}{the prior weight for edges that were not found in the Gerstein network.}
}
\description{
Function to build a prior automatically using a previously built Network on ChIP-Seq.
}
