% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfs_human.R
\name{get_tfs_human}
\alias{get_tfs_human}
\title{Function to automatically assign Human TFs given a list of Symbols.}
\usage{
get_tfs_human(gen_names)
}
\arguments{
\item{gen_names}{a vector or matrix with the Symbol Gene Names of the system to be analysed.}
}
\description{
Function to automatically assign Human TFs given a list of Symbols.
}
