% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_uni_boosting.R
\name{greedy_uni_boosting}
\alias{greedy_uni_boosting}
\title{Function to perform boosting iterations in KBoost.}
\usage{
greedy_uni_boosting(f, X, TFs, kpca, v)
}
\arguments{
\item{f}{an NxG matrix with the predicted gene expression values.}

\item{X}{an NxG matrix with the expression values.}

\item{TFs}{a matrix with integers with the columns of X that are TFs.}

\item{kpca}{a lits with the Kernel Principal Components.}

\item{v}{a double between 0 and 1 that corresponds to the shrinkage parameter.}
}
\value{
list with log likelihoods and residuals
}
\description{
Function to perform boosting iterations in KBoost.
}
