% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human_TFs-data.R
\docType{data}
\name{Human_TFs}
\alias{Human_TFs}
\title{Index of human genes' Symbols, Entrez and Ensembl for K-Boost Pacakge that correspond to transcription factors.}
\format{
matrix
}
\usage{
Human_TFs
}
\description{
Table with three columns corresponding to Symbol
}
\examples{

data(Human_TFs)
}
\references{
Lambert, S.A., et al. The Human Transcription Factors. Cell 2018;172(4):650-665.
}
\keyword{datasets}
