% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prior.R
\name{init_prior}
\alias{init_prior}
\title{Function to initialize the log prior of a model given our assumptions.}
\usage{
init_prior(prior_weights)
}
\arguments{
\item{prior_weights}{an GxK numeric matrix with the prior probability of a TF in our set to regulate a gene.}
}
\description{
Function to initialize the log prior of a model given our assumptions.
}
