% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irma_check.R
\name{irma_check}
\alias{irma_check}
\title{Function to produce the AUPR and AUROC Results on the DREAM4 Multifactorial Challenge.}
\usage{
irma_check(v, g, ite)
}
\arguments{
\item{v}{a number between 0 and 1 that is the shrinkage parameter}

\item{g}{a number larger than 0 that is the width parameter for the RBF Kernel}

\item{ite}{an integer with number of iterations.}
}
\description{
Function to produce the AUPR and AUROC Results on the DREAM4 Multifactorial Challenge.
}
