% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRMA_Gold-data.R
\docType{data}
\name{IRMA_Gold}
\alias{IRMA_Gold}
\title{IRMA Gold Standard Network}
\format{
matrix
}
\source{
https://www.sciencedirect.com/science/article/pii/S0092867409001561
}
\usage{
IRMA_Gold
}
\description{
Matrix where the rows are genes and columns are transcription factor.
}
\examples{

data(IRMA_Gold)
}
\references{
Cantone, I., et al. A Yeast Synthetic Network for In Vivo Assessment of Reverse-Engineering and Modeling Approaches. Cell 2009;137(1):172-181.
}
\keyword{datasets}
