% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irma_on-data.R
\docType{data}
\name{irma_on}
\alias{irma_on}
\title{IRMA On Dataset}
\format{
matrix
}
\source{
https://www.sciencedirect.com/science/article/pii/S0092867409001561
}
\usage{
irma_on
}
\description{
Matrix where the rows are experiments and columns are genes for the IRMA Off dataset.
}
\examples{

data(irma_on)
}
\references{
Cantone, I., et al. A Yeast Synthetic Network for In Vivo Assessment of Reverse-Engineering and Modeling Approaches. Cell 2009;137(1):172-181.
}
\keyword{datasets}
