% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_BMA.R
\name{log_BMA}
\alias{log_BMA}
\title{Function to perform Bayesian Model Averaging with log un-normalized posteriors.}
\usage{
log_BMA(posterior_list, model_list, G, K, ite)
}
\arguments{
\item{posterior_list}{a list with GxK numeric matrix with the log-posteriors of TF regulation models.}

\item{model_list}{a list with GxK logical matrix that indicates if a TF was part of a regulation model.}

\item{G}{the number of genes.}

\item{K}{the number of TFs.}

\item{ite}{the number of iterations.}
}
\value{
the Byesian model averaging result.
}
\description{
Function to perform Bayesian Model Averaging with log un-normalized posteriors.
}
