% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Function to calculate the Log-Marginal Likelihood of a Function of the Kernel PC regression model.}
\usage{
log_lik(Y, F)
}
\arguments{
\item{Y}{an NxG numeric matrix with the expression values of G genes and N observations.}

\item{F}{an NxG numeric matrix with the predicted expression values of G genes and N observations.}
}
\value{
the log marginal likelihood.
}
\description{
Function to calculate the Log-Marginal Likelihood of a Function of the Kernel PC regression model.
}
