% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_refine.R
\name{net_refine}
\alias{net_refine}
\title{Function to do a heuristic post-processing that improves accuracy. Each column is multiplied by its variance.}
\usage{
net_refine(Net)
}
\arguments{
\item{Net}{a GRN with TFs in the columns.}
}
\value{
the network with Slavek and Arodz heuristic
examples
Net =rbeta(10000,1,2)
Net = matrix(Net,100,100)
net_ref = net_refine(Net)
}
\description{
Function to do a heuristic post-processing that improves accuracy. Each column is multiplied by its variance.
}
