% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_summary_bin.R
\name{net_summary_bin}
\alias{net_summary_bin}
\title{Function to summarize the GRN filtered with a threshold,
the function produces a table version of the GRN, outdegree and indegree centrality, Katz centrality and closeness centrality.}
\usage{
net_summary_bin(GRN, TFs, thr, a, b)
}
\arguments{
\item{GRN}{An inferred networks with the predictive probabilities that a transcription facor regulates a gene.}

\item{TFs}{A vector with indexes of the rows of GRN which correspond to TFs.}

\item{thr}{a scalar between 0 and 1, edges with posterior probabilities lower than thr will be discarded.}

\item{a}{a scalar for the Katz and PageRank centrality measures. Default the inverse of the largest eigenvalue of GRN.}

\item{b}{a scalar for the Katz and PageRank centrality measures. Default is 1.}
}
\value{
a list with the network analysed.
}
\description{
Function to summarize the GRN filtered with a threshold,
the function produces a table version of the GRN, outdegree and indegree centrality, Katz centrality and closeness centrality.
}
\examples{
data(D4_multi_1)
Net = kboost(D4_multi_1)
Net_Summary = net_summary_bin(Net$GRN)

}
