% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ort_reg.R
\name{ort_reg}
\alias{ort_reg}
\title{Function to calculate a regression when regressors are orthogonal and have norm 1.}
\usage{
ort_reg(X, Y, v)
}
\arguments{
\item{X}{an NxT numeric matrix with the orthogonal normalized regressors (e.g. KPCs)}

\item{Y}{an NxG numeric matrix with the expression values of G genes and N obersvations.}

\item{v}{a number between 0 and 1 it corresponds to the shrinkage parameter.}
}
\description{
Function to calculate a regression when regressors are orthogonal and have norm 1.
}
