% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_model.R
\name{prior_model}
\alias{prior_model}
\title{Function to calculate the log prior of a model given our assumptions.}
\usage{
prior_model(prior_weights, model, prior)
}
\arguments{
\item{prior_weights}{an GxK numeric matrix with the prior probability of a TF in our set to regulate a gene.}

\item{model}{an GxK binary matrix with values of True and False if a TF was added to a model.}

\item{prior}{a Gx1 numeric matrix with the prior probability of the current model.}
}
\value{
prior probability of a model
}
\description{
Function to calculate the log prior of a model given our assumptions.
}
