% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBF_K.R
\name{RBF_K}
\alias{RBF_K}
\title{Function to calculate the RBF Kernel of a matrix X with width g.}
\usage{
RBF_K(x, g)
}
\arguments{
\item{x}{an Nx1 numeric matrix with N observations.}

\item{g}{a positive scalar with the width parameter.}
}
\value{
the matrix with the RBF kernel
}
\description{
Function to calculate the RBF Kernel of a matrix X with width g.
}
\examples{
x = rnorm(100,0,1)
k = RBF_K(x,40)

}
