% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_kpc_reg.R
\name{tf_kpc_reg}
\alias{tf_kpc_reg}
\title{Function to perform a univariate RBF KPC regression per TF in X.}
\usage{
tf_kpc_reg(X, TFs, g, v)
}
\arguments{
\item{X}{an NxG numeric matrix with the expression values of G genes and N obersvations.}

\item{TFs}{a Kx1 numeric matrix with integers that are the indexes of columns of X with genes that encode TFs.}

\item{g}{a positive scalar with the width parameter for the RBF kernel.}

\item{v}{a double between 0 and 1 with the shrinkage parameter.}
}
\description{
Function to perform a univariate RBF KPC regression per TF in X.
}
