% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_GRN_D4.R
\name{write_GRN_D4}
\alias{write_GRN_D4}
\title{Function to write output in DREAM4 Challenge Format.}
\usage{
write_GRN_D4(GRN, TFs, filename)
}
\arguments{
\item{GRN}{a GxK gene regulatory network.}

\item{TFs}{a K set of indixes of G that are TFs.}

\item{filename}{a string with the filename.}
}
\description{
Function to write output in DREAM4 Challenge Format.
}
