% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUPR_AUROC_matrix.R
\name{AUPR_AUROC_matrix}
\alias{AUPR_AUROC_matrix}
\title{Function to calculate the AUROC and AUPR of a known network.This function was made to test the R implementation of the KBoost Package.}
\usage{
AUPR_AUROC_matrix(Net, G_mat, auto_remove, TFs, upper_limit)
}
\arguments{
\item{Net}{An inferred networks with the predictive probabilities that a transcription factor regulates a gene.}

\item{G_mat}{A matrix with the gold standard network.}

\item{auto_remove}{TRUE if the auto-regulation is to be discarded.}

\item{TFs}{the indexes of the rows of Net that are TFs.}

\item{upper_limit}{Top number of edges to use. In DREAM5, the authors suggest using the top 100,000 edges only.}
}
\description{
Function to calculate the AUROC and AUPR of a known network.This function was made to test the R implementation of the KBoost Package.
}
