% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KBoost_human_symbol.R
\name{KBoost_human_symbol}
\alias{KBoost_human_symbol}
\title{Function to automatically build and run KBoost on data with human gene expression values and Symbol gene names.}
\usage{
KBoost_human_symbol(X, gen_names, g, v, ite, pos_weight, neg_weight)
}
\arguments{
\item{X}{an NxG numeric matrix with the expression values of G genes and N obersvations. The gen_names can come as column names.}

\item{gen_names}{a set of SYMBOL gene names that correspond to the names of the columns of X.}

\item{g}{a positive scalar with the width parameter for the RBF kernel. (default g = 40)}

\item{v}{a double between 0 and 1 with the shrinkage parameter. (default v =0.1)}

\item{ite}{an integer with the number of iterations (default ite = 3)}

\item{pos_weight}{a scalar between 0 and 1. This is the prior probability of observing a TF regulate a gene given that this interaction was observed before.}

\item{neg_weight}{a scalar between 0 and 1. This is the prior probability of observing a TF regulate a gene given that this interaction was NOT observed before.}
}
\value{
list with results of KBoost on a dataset with Symbol gene names.
}
\description{
Function to automatically build and run KBoost on data with human gene expression values and Symbol gene names.
}
\examples{
X = rnorm(50,0,1)
X = matrix(X,10,5)
gen_names = c("TP53","YY1","CTCF","MDM2","ESR1")
grn = KBoost_human_symbol(X,gen_names,pos_weight = 0.6, neg_weight =0.4)
}
