% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{kernel_normal}
\alias{kernel_normal}
\title{A function to perform feature normalization in kernel space.}
\usage{
kernel_normal(K)
}
\arguments{
\item{K}{an NxN numeric matrix with the kernel function with N observations.}
}
\description{
A function to perform feature normalization in kernel space.
}
