% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_matrix_D4.R
\name{tab_2_matrix_D4}
\alias{tab_2_matrix_D4}
\title{Function to produce the gold standard of the DREAM4 Multifactorial Challenge in matrix format.}
\usage{
tab_2_matrix_D4(g_table, G)
}
\arguments{
\item{g_table}{the network in table format. The first column is the Tf, the second column the gene, and the third indicates if there is an interaction.}

\item{G}{the number of genes.}
}
\description{
Function to produce the gold standard of the DREAM4 Multifactorial Challenge in matrix format.
}
